refactor: last edit for demo

+++ App.js
@@ -1,5 +1,7 @@
 import React from 'react';
 import Icon from 'react-native-vector-icons/Ionicons';
+import { Alert } from 'react-native';
+
 import { deleteUserCar, deleteCustomFieldCar } from './api/UserCar';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
@@ -28,7 +30,10 @@ const SettingStack = createStackNavigator();
 const MainStack = createBottomTabNavigator();
 
 const AuthStackScreens = () => (
-  <AuthStack.Navigator>
+  <AuthStack.Navigator screenOptions={{
+    headerTitleAlign: 'center',
+    headerTintColor: 'black',
+  }}>
     <AuthStack.Screen name="Login" options={{ title: strings.mainStack.Login }}>
       {props => <LoginScreen {...props} />}
     </AuthStack.Screen>
@@ -38,15 +43,40 @@ const AuthStackScreens = () => (
 );
 
 const SettingStackScreens = () => (
-  <SettingStack.Navigator>
+  <SettingStack.Navigator screenOptions={{
+    headerTitleAlign: 'center',
+    headerTintColor: 'black',
+  }}>
     <SettingStack.Screen name="Setting" component={SettingScreen} options={{ title: strings.mainStack.Setting }} />
     <SettingStack.Screen name="EditProfile" component={EditProfileScreen} options={{ title: strings.mainStack.EditProfile }} />
     <SettingStack.Screen name="ChangePassword" component={ChangePasswordScreen} options={{ title: strings.mainStack.ChangePassword }} />
   </SettingStack.Navigator>
 );
 
+const confirmDeletion = (title, message, onDelete) => {
+  Alert.alert(
+    title,
+    message,
+    [
+      {
+        text: strings.mainStack.noMessage,
+        onPress: () => console.log('Cancel Pressed'),
+        style: 'cancel',
+      },
+      {
+        text: strings.mainStack.yesMessage,
+        onPress: onDelete,
+      },
+    ],
+    { cancelable: false },
+  );
+};
+
 const CarStackScreens = () => (
-  <CarStack.Navigator>
+  <CarStack.Navigator screenOptions={{
+    headerTitleAlign: 'center',
+    headerTintColor: 'black',
+  }}>
     <CarStack.Screen name="Home" component={CarsListScreen} options={({ navigation }) => ({
       title: strings.mainStack.Home,
       headerRight: () => (
@@ -62,24 +92,36 @@ const CarStackScreens = () => (
     <CarStack.Screen name="CustomField" component={CustomFieldScreen} options={({ route, navigation }) => ({
       title: route.params.customField?.name || strings.mainStack.CustomFieldList,
       headerRight: () => route.params.customField && (
-        <Icon.Button name="trash" size={24} color="red" backgroundColor="transparent" onPress={() => {
-          deleteCustomFieldCar(route.params.car.unique_key, route.params.customField.id).then(() => {
-            navigation.navigate('CustomFieldList', { refresh: true, car: route.params.car });
-          }).catch((error) => {
-            console.error('Error deleting customField:', error);
-          });
+        <Icon.Button name="trash" size={24} color="gray" backgroundColor="transparent" onPress={() => {
+          confirmDeletion(
+            strings.mainStack.deleteCustomFieldTitle,
+            strings.mainStack.deleteQuestionText,
+            () => {
+              deleteCustomFieldCar(route.params.car.unique_key, route.params.customField.id).then(() => {
+                navigation.navigate('CustomFieldList', { refresh: true, car: route.params.car });
+              }).catch((error) => {
+                console.error('Error deleting customField:', error);
+              });
+            },
+          );
         }} />
       ),
     })} />
     <CarStack.Screen name="CarScreen" component={CarScreen} options={({ route, navigation }) => ({
       title: route.params?.car?.name || 'Car',
       headerRight: () => route.params.car && (
-        <Icon.Button name="trash" size={24} color="red" backgroundColor="transparent" onPress={() => {
-          deleteUserCar(route.params.car.unique_key).then(() => {
-            navigation.navigate('Home', { refresh: true });
-          }).catch((error) => {
-            console.error('Error deleting car:', error);
-          });
+        <Icon.Button name="trash" size={24} color="gray" backgroundColor="transparent" onPress={() => {
+          confirmDeletion(
+            strings.mainStack.deleteCarTitle,
+            strings.mainStack.deleteQuestionText,
+            () => {
+              deleteUserCar(route.params.car.unique_key).then(() => {
+                navigation.navigate('Home', { refresh: true });
+              }).catch((error) => {
+                console.error('Error deleting car:', error);
+              });
+            },
+          );
         }} />
       ),
     })} />
@@ -99,25 +141,32 @@ const CarStackScreens = () => (
 );
 
 const MainStackScreens = () => (
-  <MainStack.Navigator screenOptions={({ route }) => ({
-    tabBarIcon: ({ focused, color, size }) => {
-      let iconName;
-      if (route.name === 'Home') {
-        iconName = focused ? 'car' : 'car-outline';
-      } else if (route.name === 'Setting') {
-        iconName = focused ? 'settings' : 'settings-outline';
-      } else if (route.name === 'Notification') {
-        iconName = focused ? 'notifications' : 'notifications-outline';
-      }
-      return <Ionicons name={iconName} size={size} color={color} />;
-    },
-    tabBarShowLabel: false,
-  })} tabBarOptions={{
+<MainStack.Navigator 
+  screenOptions={({ route }) => ({ 
+    tabBarIcon: ({ focused, color, size }) => { 
+      let iconName; 
+      if (route.name === 'Home') { 
+        iconName = focused ? 'car' : 'car-outline'; 
+      } else if (route.name === 'Setting') { 
+        iconName = focused ? 'settings' : 'settings-outline'; 
+      } else if (route.name === 'Notification') { 
+        iconName = focused ? 'notifications' : 'notifications-outline'; 
+      } 
+      return <Ionicons name={iconName} size={size} color={color} />; 
+    }, 
+    tabBarShowLabel: false, 
+  })}
+  tabBarOptions={{
     activeTintColor: 'blue',
     inactiveTintColor: 'gray',
-  }}>
+    style: {
+      backgroundColor: 'green',
+    }
+  }}
+>
     <MainStack.Screen name="Home" component={CarStackScreens} options={{ headerShown: false }} />
-    <MainStack.Screen name="Notification" component={NotificationScreen} options={{ headerTitle: strings.mainStack.Notification }} />
+    <MainStack.Screen  name="Notification" component={NotificationScreen} options={{ headerTitle: strings.mainStack.Notification  ,   headerTitleAlign: 'center'
+}} />
     <MainStack.Screen name="Setting" component={SettingStackScreens} options={{ headerShown: false }} />
   </MainStack.Navigator>
 );
